refactor(store-core): extract file-op result handlers into handlers/file.ts (audit P2-3)
Second slice of the P2-3 handlers/index.ts split. Move the four
file-operation result handlers (directory_listing / file_listing /
file_content / write_file_result), their payload interfaces, and the
private extractEntriesPayload helper into a new handlers/file.ts.

Pure move, no logic change: the extracted block is byte-identical to the
removed lines (proven by diff). The block is fully self-contained — it uses
no shared barrel helpers (resolveSessionId etc.) and no ../types imports
(concrete entry types live downstream in app/dashboard). The barrel
re-exports via `export * from './file'`, so the public surface is unchanged
and every consumer is unaffected.

handlers/index.ts: 5399 -> 5252 lines.
